feat: Add window_resize_listener widget
Useful for flap / leaflet and maybe other stuff in the future

The widget is rebuilt on every resize but there is no way to tell if it
was caused by a window resize or some other state change. This way we
can only update the needed parts on a window resize.