Don't check copy paste, remove dry-run in publish
The copy paste checker was triggered for dart constructors in class such as AdwButton which is counter-productive. I've also removed the dry-run from the publish CI.

I'm sorry. I know I've introduce the super-linter, I thought it would be good but it's more a headache in the end since most of the linter is used in ci.yaml. I still think it's useful for the Github secret check, especially once we'll publish release using the publish CI, but if you thik it's too annoying, I really don't mind removing it.